requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
libretranslatepy==2.1.1
lxml==5.2.2
mysqlclient==2.2.4
requests==2.32.3
sqlparse==0.5.1
translate==3.6.1
tzdata==2024.1
urllib3==2.2.2
